requirements.txt
# Python Dependencies for Churn Model ML Pipeline
# Oracle Machine Learning for Python (OML4Py)
# Note: OML4Py is typically installed via Oracle client, not pip
# Install Oracle Instant Client and OML4Py separately

# Core Data Science Libraries
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0

# Oracle Database Connection (for Python scripts)
# Use oracledb (recommended) or cx_Oracle
oracledb>=2.0.0

# Model Training and Evaluation
xgboost>=2.0.0

# Utilities
python-dotenv>=1.0.0

# Optional: For data visualization and analysis
matplotlib>=3.7.0
seaborn>=0.12.0

# Optional: For Kaggle dataset download
kaggle>=1.5.0
